+++ src/app/component-library/catalog/asset-detail-dialog/asset-detail-dialog.service.ts
@@ -22,7 +22,7 @@ export class AssetDetailDialogService {
     return showDialogUntil(
       this.dialog,
       AssetDetailDialogComponent,
-      {data, maxHeight: '90vh'},
+      {data, maxWidth: '1000px', maxHeight: '90vh', autoFocus: false},
       until$,
     );
   }