+++ src/app/core/services/api/fake-backend/connector-fake-impl/data/test-assets.ts
@@ -1,12 +1,95 @@
 import {UiAsset} from '@sovity.de/edc-client';
 
 export namespace TestAssets {
+  const markdownDescription = `Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod
+  tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At
+  vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren,
+  no sea takimata sanctus est Lorem ipsum dolor sit amet.
+
+  ![scenery2](https://images.pexels.com/photos/255419/pexels-photo-255419.jpeg?cs=srgb&dl=pexels-pixabay-255419.jpg&fm=jpg)
+  
+  Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod
+  tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At
+  vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren,
+
+  Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod
+  tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At
+  vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren,
+
+  Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod
+  tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At
+  vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren,
+  
+  Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod
+  tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At
+  vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren,
+
+ ![scenery](https://images.rawpixel.com/image_800/cHJpdmF0ZS9sci9pbWFnZXMvd2Vic2l0ZS8yMDIyLTA1L3NrOTc5MS1pbWFnZS1rd3Z1amE5Ni5qcGc.jpg) Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren,
+
+
+
+
+  # Omen
+  
+  This is **bold!** This is _italic_. This is inline \`code\`.
+  
+  > here we quote
+
+  Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod
+  tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At
+  vero eos et accusam et justo duo dolores et ea rebum.
+  
+  ## Sage
+  
+  - list item 1
+  - list item 2
+  
+  ### Raze
+  
+  1. list item 1
+  2. list item 2
+  
+  #### Cypher
+  
+  \`\`\`javascript
+  alert(1);
+  \`\`\`
+  
+  ##### Jett
+  
+  [Sovity Website](https://sovity.de/)
+  
+  **Table**
+| Item              | In Stock | Price | Description |
+| :---------------- | :------: | ----: | :---------- |
+| Python Hat        |   True   | 24.99 | This is a long description to test the scrolling behavior of the table. This is a long description to test the scrolling behavior of the table. |
+| SQL Hat           |   True   | 24.99 | This is a long description to test the scrolling behavior of the table. This is a long description to test the scrolling behavior of the table. |
+| Codecademy Tee    |  False   | 20.99 | This is a long description to test the scrolling behavior of the table. This is a long description to test the scrolling behavior of the table. |
+| Codecademy Hoodie |  False   | 43.99 | This is a long description to test the scrolling behavior of the table. This is a long description to test the scrolling behavior of the table. |
+`;
+
+  const shortMarkdownDescription = `# Short Description
+
+This is a short description text that should be fully rendered without being **collapsed**. No *show more* button should be visible.
+`;
+
   export const boring: UiAsset = {
     assetId: 'data-sample-ckd-skd-demands-2023-Jan',
     title: 'data-sample-ckd-skd-demands-2023-Jan',
     connectorEndpoint: 'https://my-other-connector/api/dsp',
     participantId: 'MDSL1234XX.C1234XX',
     creatorOrganizationName: 'my-other-connector',
+  };
+
+  export const short: UiAsset = {
+    assetId: 'data-sample-ckd-skd-demands-2023-Feb',
+    title: 'data-sample-ckd-skd-demands-2023-Feb',
+    connectorEndpoint: 'https://my-other-connector/api/dsp',
+    participantId: 'MDSL1234XX.C1234XX',
+    creatorOrganizationName: 'my-other-connector',
+    description: shortMarkdownDescription,
+    descriptionShortText:
+      'Short Description This is a short description text that should be fully rendered without being collapsed. No show more button should be visible.',
   };
 
   export const full: UiAsset = {
@@ -18,8 +101,9 @@ export namespace TestAssets {
     creatorOrganizationName: 'My-German-OEM',
     keywords: ['automotive', 'part-demands', '2023', 'January'],
     mediaType: 'application/json',
-    description:
-      'Part demands for CKD/SKD parts for January 2023. Split by plant / day / model code.',
+    description: markdownDescription,
+    descriptionShortText:
+      'Part demands for CKD/SKD parts January 2023 Split by plant / day / model code. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum.',
     language: 'https://w3id.org/idsa/code/EN',
     publisherHomepage:
       'https://teamabc.departmentxyz.my-german-oem.de/offers/ckd-skd-demands',